add full function of department

+++ src/main/java/Department.java
@@ -61,18 +61,32 @@ public class Department {
         //Get data from keyboard
         String dp_name = this.department_name;
         //Write SQL add data
-        String sql = "INSERT INTO departments(department_name) VALUES (" + dp_name + ")";
+        String sql = "INSERT INTO departments(department_name) VALUES ('" + dp_name + "')";
         //Run sql
         Statement statement = connection.createStatement();
         statement.executeUpdate(sql);
         //Close SQL
         DatabaseConnection.closeConnection(connection);
+        System.out.println("Add completed!");
     }
 
     //metod update data
     public void updateDP(int id) throws SQLException {
         //Connect Db
         Connection connection = DatabaseConnection.openConnection();
+        //Get uid department before delete
+        String checkUid = "SELECT id FROM departments ORDER BY id DESC";
+        //Run sql get uid
+        Statement statementUid = connection.createStatement();
+        ResultSet results = statementUid.executeQuery(checkUid);
+        //Get uid
+        results.next();
+        int uid = results.getInt("id");
+        //Check uid employee
+        if (id > uid) {
+            System.out.println("ID not found! try again!");
+            return;
+        }
         //Get data input
         String dp_name = this.department_name;
         //Write a query update data with id
@@ -82,12 +96,26 @@ public class Department {
         statement.executeUpdate(sql);
         //Close SQL
         DatabaseConnection.closeConnection(connection);
+        System.out.println("Update completed!");
     }
 
     //method delete data in db
     public void deleteDP(int id) throws SQLException {
         //Connect Db
         Connection connection = DatabaseConnection.openConnection();
+        //Get uid department before delete
+        String checkUid = "SELECT id FROM departments ORDER BY id DESC";
+        //Run sql get uid
+        Statement statementUid = connection.createStatement();
+        ResultSet results = statementUid.executeQuery(checkUid);
+        //Get uid
+        results.next();
+        int uid = results.getInt("id");
+        //Check uid employee
+        if (id > uid) {
+            System.out.println("ID not found! try again!");
+            return;
+        }
         //Write a query update data with id
         String sql = "DELETE FROM departments WHERE id = " + id;
         //Run query
@@ -95,24 +123,6 @@ public class Department {
         statement.executeUpdate(sql);
         //Close SQL
         DatabaseConnection.closeConnection(connection);
-    }
-
-    //method search by id
-    public void searchbyId(int id) throws SQLException {
-        //Connect DB
-        Connection connection = DatabaseConnection.openConnection();
-        //Write query to search by id
-        String sql = "SELECT * FROM Departments WHERE id = " + id;
-        //Run query
-        Statement statement = connection.createStatement();
-        ResultSet results = statement.executeQuery(sql);
-        //Display data was given
-        while (results.next()){
-            //get department_name
-            String department_name = results.getString("department_name");
-            //display
-            System.out.println("ID: " + id + ", department_name: " + department_name);
-            //close
-        }
+        System.out.println("Delete completed!");
     }
 }
